@@ -1,3 +1,4 @@
+import 'package:flutter/cupertino.dart';
 import 'package:square_counter/src/errors/errors.dart';
 import 'package:flutter_triple/flutter_triple.dart';
 
@@ -6,13 +7,22 @@ import 'square_store.dart';
 class HomeStore extends StreamStore<List<SquareStore>, SquareError> {
   HomeStore() : super([]);
 
-  void addSquare() {
+  void initializeSquare(List<SquareStore> squares) {
+    setState(squares);
+  }
+
+  addSquare() async {
+    setLoading(true);
+    await Future.delayed(Duration(seconds: 1));
+
     if (state.length < 9) {
-      final newList = List<SquareStore>.from(state)..add(SquareStore(this, index: state.length + 1));
+      final newList = List<SquareStore>.from(state)
+        ..add(SquareStore(this, index: state.length + 1));
       setState(newList);
     } else {
       setError(SquareError('Limite de squares atingido!'));
     }
+    setLoading(false);
   }
 
   void removeSquare() async {
